Extract modals into components
Refactored the profile page a little bit by moving add, edit, and delete modal logic into separate, reusable ActForm and ActDelete components.
